+++ lib/omega_architecture.dart
@@ -16,6 +16,7 @@ export 'omega/core/channel/omega_channel.dart';
 export 'omega/core/semantics/omega_intent.dart';
 export 'omega/core/semantics/omega_intent_name.dart';
 export 'omega/core/semantics/omega_event_name.dart';
+export 'omega/core/semantics/omega_typed_event.dart';
 export 'omega/core/events/omega_event.dart';
 export 'omega/core/types/omega_failure.dart';
 export 'omega/core/types/omega_object.dart';